Add files via upload
First mini-project using Cisco Packet Tracer and creating and configuring a SOHO network using a router, 2 switches, and some PCs. Connecting the devices using a wired connection and assigning IP addresses to the network. The network address is subnetted into two networks and the connectivity between the two subnets is verified using the PING command.